refactor: 칼로리 데이터 형식 통일

+++ lib/features/workout_history/data/models/apple_health_workout_model.dart
@@ -19,7 +19,7 @@ class AppleHealthWorkoutModel {
   final String endTime;
   final int duration;
   final double distance;
-  final int calories;
+  final double calories;
   final String source;
   final String title;
   final List<HeartRateDataModel> heartRateData;
